Makefile: the `base64` test util was not building correctly
Only the codecs in the libbase64.a library were being built with the
proper architecture flags. The codecs that were being built into the
base64 test utility were being built without any architecture-specific
flags, causing them to fail at runtime.

Overhaul the Makefile so that both the test utility and the library
depend on the same set of object files. Ensure the object files are
always built with the proper architecture flags.